+++ src/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"goji/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
